Refactor front-end building process
