+++ website/prereg/views.py
@@ -9,27 +9,10 @@ Other resources that are a part of the Prereg Challenge:
 * website/static/css/prereg.css
 """
 from flask import request
-from modularodm import Q
 
 from framework.auth import decorators
 from framework.utils import iso8601format
-from website.util import permissions
-from website.prereg.utils import get_prereg_schema, serialize_campaign_context
-
-def drafts_for_user(user, campaign):
-    from website import models  # noqa
-
-    user_projects = models.Node.find(
-        Q('is_deleted', 'eq', False) &
-        Q('permissions.{0}'.format(user._id), 'in', [permissions.ADMIN])
-    )
-    PREREG_CHALLENGE_METASCHEMA = get_prereg_schema(campaign)
-    return models.DraftRegistration.find(
-        Q('registration_schema', 'eq', PREREG_CHALLENGE_METASCHEMA) &
-        Q('approval', 'eq', None) &
-        Q('registered_node', 'eq', None) &
-        Q('branched_from', 'in', [p._id for p in user_projects])
-    )
+from website.prereg import utils
 
 @decorators.must_be_logged_in
 def prereg_landing_page(auth, **kwargs):
@@ -41,20 +24,20 @@ def prereg_landing_page(auth, **kwargs):
         if node.has_permission(user=auth.user, permission='admin')
     ]
     has_projects = bool(registerable_nodes)
-    has_draft_registrations = bool(drafts_for_user(auth.user, campaign).count())
+    has_draft_registrations = bool(utils.drafts_for_user(auth.user, campaign).count())
 
-    ret = {
+    return {
         'has_draft_registrations': has_draft_registrations,
         'has_projects': has_projects,
+        'campaign_long': utils.PREREG_CAMPAIGNS[campaign],
+        'campaign_short': campaign
     }
-    ret.update(serialize_campaign_context(campaign))
-    return ret
 
 @decorators.must_be_logged_in
 def prereg_draft_registrations(auth, **kwargs):
     """API endpoint; returns prereg draft registrations the user can resume"""
     campaign = kwargs.get('campaign', 'prereg')
-    drafts = drafts_for_user(auth.user, campaign)
+    drafts = utils.drafts_for_user(auth.user, campaign)
     return {
         'draftRegistrations': [
             {
